Name the version 'ubu12.04' instead of 'precise'.
If we release-upgrade, the increasing version number makes sure we get the
correct package. (Yes, for Ubuntu they still increase, but after Yakkety and
Zesty we'll probably end up with an A and then the ordering stops.)

While 3.5...~precise looks better than 3.5...~ubu12.04, I'll have to go with
consistency and ease of upgrading here.